requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
docopt==0.6.2
fastapi==0.73.0
h11==0.14.0
idna==3.4
motor==2.5.1
pipreqs==0.4.11
pydantic==1.10.2
pymongo==3.13.0
python-decouple==3.6
requests==2.28.1
sniffio==1.3.0
starlette==0.17.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn
yarg==0.1.9
